requirements.txt
aliyun-python-sdk-core-v3==2.9.3
aliyun-python-sdk-dysmsapi==1.0.0
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==2.0
django-debug-toolbar==1.10.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.0
idna==2.7
mysqlclient==1.3.13
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
python-memcached==1.59
pytz==2018.5
qiniu==7.2.2
requests==2.20.1
shortuuid==0.5.0
six==1.11.0
sqlparse==0.2.4
urllib3==1.24.1
